divided the chessviz file into different libraries and scattered it into folders, made a makefile, brought it to a working state
