+++ .gitignore
@@ -1,3 +1,7 @@
 .*
 a.*
 *.o
+*.a
+*.exe
+*.out
+chessviz
