+++ README.md
@@ -1,0 +1,1 @@
+Chess. Just chess. Not absolutely complete, but chess. Written as part of laboratory work on TRPO by Rylov Lev.
