+++ src/chessviz/int_switch_char.h
@@ -1,0 +1,4 @@
+#pragma once
+
+int char_int(char n);
+char int_char(int n);
